fix: Various linter fixes and updates (#53)
* fix: grept apply

* chore: Update outputs.tf with additional outputs

This commit adds three new outputs to the outputs.tf file: resource_id, name, and system_assigned_mi_principal_id. These outputs provide information about the Databricks Workspace in the Azure management plane. The resource_id output contains the ID of the workspace, the name output contains the name of the workspace, and the system_assigned_mi_principal_id output contains the principal ID of the system assigned managed identity.

Co-authored-by: Sebastian Graef <segraef@microsoft.com>

* chore: Add location variable to variables.tf

This commit adds a new variable "location" to the variables.tf file. The "location" variable specifies the Azure region where the resource should be deployed. It is set as non-nullable.

Co-authored-by: Sebastian Graef <segraef@microsoft.com>

* feat: Add location variable to variables.tf

This commit adds a new variable "location" to the variables.tf file. The "location" variable specifies the Azure region where the resource should be deployed. It is set as non-nullable.

* feat: Add location variable to variables.tf

* chore: Remove system_assigned_mi_principal_id output from outputs.tf

* chore: Update locals.tf to fix resource_group_location

The code change in locals.tf fixes the reference to the location of the parent resource group. The previous code used an incorrect index, which has been corrected to use the correct reference. This change ensures that the resource_group_location variable is set correctly.

Co-authored-by: Sebastian Graef <segraef@microsoft.com>

* feat: Add location variable to module databricks

This commit adds a new variable "location" to the module databricks in the main.tf file. The "location" variable specifies the Azure region where the resource should be deployed. It is set based on the randomly selected region from the module.regions output.

* feat: Add location variable to module databricks

* feat: Add location variable to module databricks

* feat: Update module versions in main.tf

This commit updates the versions of the modules used in the main.tf file to ">= 0.8.0". This ensures that the latest stable versions of the modules are being used.

* feat: Update module versions to >= 0.8.0

This commit updates the versions of the modules used in the main.tf and examples/*.tf files to ">= 0.8.0". This ensures that the latest stable versions of the modules are being used.

* Update module versions to >= 0.4.1

* ```text
feat: Update azurerm_management_lock count logic in main.tf

This commit updates the count logic for the azurerm_management_lock resource in the main.tf file. The previous logic checked if the var.lock.kind was not equal to "None", but it has been updated to check if var.lock.kind is not null. This change ensures that the azurerm_management_lock resource is created when the lock kind is specified.

* feat: Update azurerm_management_lock count logic in main.tf

* feat: Update azurerm_management_lock count logic in main.tf

* feat: Update azurerm_management_lock count logic in main.tf

* feat: Update azurerm_management_lock count logic in main.tf

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>